chore: 迁移更新地址的代码到 SignalsExtenesions 中
